refactor: задание на проверку.
