+++ src/main/java/ru/practicum/shareit/booking/BookingMapper.java
@@ -3,9 +3,8 @@ package ru.practicum.shareit.booking;
 import ru.practicum.shareit.booking.dto.BookingDto;
 
 public class BookingMapper {
-
-        public BookingDto BookingToBookingDto(Booking booking) {
-            return new BookingDto(booking);
-        }
+    public BookingDto bookingToBookingDto(Booking booking) {
+        return new BookingDto(booking);
+    }
 
 }
