+++ src/main/java/ru/practicum/shareit/exceptions/BadRequestException.java
@@ -1,6 +1,6 @@
 package ru.practicum.shareit.exceptions;
 
-public class BadRequestException extends RuntimeException{
+public class BadRequestException extends RuntimeException {
     public BadRequestException(String message) {
         super(message);
     }
