+++ src/main/java/ru/practicum/shareit/item/ItemMapper.java
@@ -9,16 +9,17 @@ import java.util.List;
 public class ItemMapper {
 
     public ItemDto itemToItemDto(Item item) {
-       return new ItemDto(item, getCountRent(item));
+        return new ItemDto(item, getCountRent(item));
     }
 
     public List<ItemDto> getItemDtoList(List<Item> items) {
         List<ItemDto> itemDtoList = new ArrayList<>();
-        for(Item item : items) {
+        for (Item item : items) {
             itemDtoList.add(itemToItemDto(item));
         }
         return itemDtoList;
     }
+
     private Integer getCountRent(Item item) {
         return null;
     }
