+++ src/main/java/ru/practicum/shareit/request/RequestMapper.java
@@ -3,8 +3,7 @@ package ru.practicum.shareit.request;
 import ru.practicum.shareit.request.dto.ItemRequestDto;
 
 public class RequestMapper {
-
-    public ItemRequestDto ItemRequestToItemRequestDto(ItemRequest itemRequest) {
+    public ItemRequestDto itemRequestToItemRequestDto(ItemRequest itemRequest) {
         return new ItemRequestDto(itemRequest);
     }
