+++ src/main/java/ru/practicum/shareit/user/dao/UserDao.java
@@ -1,6 +1,5 @@
 package ru.practicum.shareit.user.dao;
 
-import ru.practicum.shareit.item.model.Item;
 import ru.practicum.shareit.user.User;
 
 import java.util.HashMap;
@@ -17,8 +16,6 @@ public interface UserDao {
     User remove(Long id);
 
     User get(Long id);
-
-  //  void addItem(User user, Item item);
 
     boolean containsKey(Long id);
