--- src/main/java/ru/practicum/shareit/validator/Validator.java
@@ -1,21 +1,0 @@
-package ru.practicum.shareit.validator;
-
-import ru.practicum.shareit.exceptions.UserIsPresentException;
-import ru.practicum.shareit.user.User;
-
-public class Validator {
-
-
-//    private void emailIsPresent(User user) {
-//        users.values()
-//                .stream()
-//                .filter(x -> x.getEmail().equals(user.getEmail()))
-//                .findFirst()
-//                .ifPresent(u -> {
-//                    log.warn("User with email: {} is already exists.", user.getEmail());
-//                    throw new UserIsPresentException("User with this email is already exists.");
-//                });
-//    }
-
-
-}
